completly removed DOM
